gen v2: Make generic JavaScript template work
