+++ gen/javascript/client_helpers.js.tmpl
