+++ gen/javascript/proto.gen.js.tmpl
